+++ e2e-test/cypress/integration/project_history_spec.js
@@ -18,15 +18,15 @@ describe("Project's history", function() {
     cy.get("#project-history-button").click();
 
     // Count history items => should be one
-    cy.get("#history-list li.history-item")
+    cy.get("[data-test=history-list] li.history-item")
       .first()
       .should("be.visible");
-    cy.get("#history-list")
+    cy.get("[data-test=history-list]")
       .find("li.history-item")
       .should("have.length", 1);
 
     // Make sure it's a creation event
-    cy.get("#history-list")
+    cy.get("[data-test=history-list]")
       .find("li.history-item")
       .first()
       .should("contain", "created project");
@@ -43,21 +43,21 @@ describe("Project's history", function() {
     cy.get("#project-history-button").click();
 
     // Count history items => should be two
-    cy.get("#history-list li.history-item")
+    cy.get("[data-test=history-list] li.history-item")
       .first()
       .should("be.visible");
-    cy.get("#history-list")
+    cy.get("[data-test=history-list]")
       .find("li.history-item")
       .should("have.length", 2);
 
     // Make sure the oldest entry is the create event
-    cy.get("#history-list")
+    cy.get("[data-test=history-list]")
       .find("li.history-item")
       .last()
       .should("contain", "created project");
 
     // Make sure the newest entry is the assign event
-    cy.get("#history-list")
+    cy.get("[data-test=history-list]")
       .find("li.history-item")
       .first()
       .should("contain", "assigned project");