+++ frontend/src/pages/Workflows/WorkflowitemHistoryTab/reducer.js
@@ -3,14 +3,16 @@ import { fromJS } from "immutable";
 import {
   SET_TOTAL_WORKFLOWITEM_HISTORY_ITEM_COUNT,
   FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE,
-  FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE_SUCCESS
+  FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE_SUCCESS,
+  RESET_WORKFLOWITEM_HISTORY
 } from "./actions";
-import { CLOSE_WORKFLOWITEM_DETAILS } from "../Workflows/actions";
+import { CLOSE_WORKFLOWITEM_DETAILS } from "../actions";
+import { LOGOUT } from "../../Login/actions";
 
 const historyPageSize = 30;
 
 const initialState = fromJS({
-  events: [],
+  historyItems: [],
   totalHistoryItemCount: 0,
   historyPageSize: historyPageSize,
   currentHistoryPage: 0,
@@ -31,12 +33,13 @@ export default function reducer(state = initialState, action) {
 
     case FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE_SUCCESS:
       return state.merge({
-        events: state.get("events").concat(fromJS(action.events).reverse()),
+        historyItems: state.get("historyItems").concat(fromJS(action.events).reverse()),
         currentHistoryPage: action.currentHistoryPage,
         isHistoryLoading: false
       });
-
+    case RESET_WORKFLOWITEM_HISTORY:
     case CLOSE_WORKFLOWITEM_DETAILS:
+    case LOGOUT:
       return initialState;
 
     default: